+++ lib/completion.ts
@@ -60,8 +60,8 @@ const blocks = [
     snippet: '##############################\n##  ${1:        Heading       }  ##\n##############################\n$2',
     displayText: '## Heading ##'
   },
-  { snippet: 'Show\n  ${1:Filter}\n  ${2:Action}' },
-  { snippet: 'Hide\n  ${1:Filter}' }
+  { snippet: 'Show\n  ${1:Rule}' },
+  { snippet: 'Hide\n  ${1:Rule}' }
 ]
 
 const filters = [
@@ -277,22 +277,12 @@ function getSuggestions(args: SuggestionParams) {
   if(topScope == 'source.poe') {
     suggestions = suggestions.concat(blocks);
   } else if(topScope == 'line.empty.poe' || topScope == 'line.unknown.poe') {
-    if(cursorScopes.indexOf('show.block.poe') != -1) {
-      if(prefix == 'Action') {
-        suggestions = suggestions.concat(actions);
-        shouldPruneSuggestions = false;
-      } else if(prefix == 'Filter') {
-        suggestions = suggestions.concat(filters);
+    if(cursorScopes.indexOf('block.poe') != -1) {
+      if(prefix == 'Rule') {
+        suggestions = suggestions.concat(actions, filters);
         shouldPruneSuggestions = false;
       } else {
         suggestions = suggestions.concat(blocks, actions, filters);
-      }
-    } else if(cursorScopes.indexOf('hide.block.poe') != -1) {
-      if(prefix == 'Filter') {
-        suggestions = suggestions.concat(filters);
-        shouldPruneSuggestions = false;
-      } else {
-        suggestions = suggestions.concat(blocks, filters);
       }
     }
   } else {